cmsis: remove startup_efm32lg.c
Because this file shares the same name as startup_efm32lg.S, the
makefiles are picking up this in favor of the assembly implementation
(because that is the first pattern that matches).

This is an issue because there is a bug in GCC [1] which causes the
section attribute to be ignored when compiled with link-time
optimization. The result is that the ISRs aren't correctly installed
into the beginning of flash.

Instead of rearranging the patterns in the makefiles, this C
implementation will be removed since it doesn't currently work with GCC.

[1]: https://bugs.launchpad.net/gcc-arm-embedded/+bug/1418073